+++ src/reconstruction.jl
@@ -1,61 +1,104 @@
 using StaticArrays
-export Reconstruction, MDReconstruction
+export reconstruct, DelayEmbedding, AbstractEmbedding
 
 #####################################################################################
-#                            Reconstruction Object                                  #
+#                        Delay Embedding Reconstruction                             #
 #####################################################################################
-abstract type AbstractReconstruction{D, T, τ} <: AbstractDataset{D, T} end
+"""
+    AbstractEmbedding
+Super-type of embedding methods. Use `subtypes(AbstractEmbedding)` for available
+methods.
+"""
+abstract type AbstractEmbedding <: Function end
 
 """
-    Reconstruction(s::AbstractVector, D, τ) <: AbstractDataset
-`D`-dimensional delay-coordinates reconstruction object with delay `τ`,
-created from a timeseries `s`.
+    DelayEmbedding(D, τ) -> `embedding`
+Return a delay coordinates embedding structure to be used as a functor,
+given a timeseries and some index. Calling
+```julia
+embedding(s, n)
+```
+will create the `n`-th reconstructed vector of the embedded space, which has `D`
+temporal neighbors with delay(s) `τ`. See [`reconstruct`](@ref) for more.
 
-Notice that the number of temporal neighbors is `D-1`!
+*Be very careful when choosing `n`, because `@inbounds` is used internally.*
+"""
+struct DelayEmbedding{D} <: AbstractEmbedding
+    # Notice that the type-parameter D here is not the number of temporal neighbors
+    # but plus one. That is because you cannot declare something like
+    # ::SVector{D+1}. It is not allowed.
+    delays::SVector{D, Int}
+end
+
+function DelayEmbedding(D, τ)
+    if typeof(τ) <: Integer
+        idxs = [k*τ for k in 0:D]
+        return DelayEmbedding{D+1}(SVector{D+1, Int}(idxs...))
+    elseif typeof(τ) <: AbstractArray{<:Integer}
+        D != length(τ) && throw(ArgumentError(
+        "Delay time vector length must equal the number of spatial neighbors."
+        ))
+        if !issorted(τ)
+            @warn "Delay times are not sorted. Sorting now."
+            τ = sort(τ)
+        end
+        return DelayEmbedding{D+1}(SVector{D+1, Int}(0, τ...))
+    end
+end
+
+@generated function (r::DelayEmbedding{D})(s::AbstractArray{T}, i) where {D, T}
+    gens = [:(s[i + r.delays[$k]]) for k=1:D]
+    quote
+        @inbounds return SVector{$D,T}($(gens...))
+    end
+end
+
+"""
+    reconstruct(s, D, τ)
+Reconstruct `s` using the delay coordinates embedding with `D` temporal neighbors
+and delay `τ` and return the result as a [`Dataset`](@ref).
 
 ## Description
-If `τ` is an integer, then the ``n``th row of a `Reconstruction`
-is
+### Single Timeseries
+If `τ` is an integer, then the ``n``-th entry of the embedded space is
 ```math
-(s(n), s(n+\\tau), s(n+2\\tau), \\dots, s(n+(D-1)\\tau))
+(s(n), s(n+\\tau), s(n+2\\tau), \\dots, s(n+D\\tau))
 ```
 If instead `τ` is a vector of integers, so that `length(τ) == D`,
-then the ``n``th row is
+then the ``n``-th entry is
 ```math
-(s(n+\\tau[1]), s(n+\\tau[2]), s(n+\\tau[3]), \\dots, s(n+\\tau[D]))
+(s(n), s(n+\\tau[1]), s(n+\\tau[2]), \\dots, s(n+\\tau[D]))
 ```
 
-The reconstruction object `R` can have same
+The reconstructed dataset can have same
 invariant quantities (like e.g. lyapunov exponents) with the original system
 that the timeseries were recorded from, for proper `D` and `τ` [1, 2].
-
 The case of different delay times allows reconstructing systems with many time scales,
 see [3].
 
-## Multi-dimensional `Reconstruction`
-To make a reconstruction out of a multi-dimensional timeseries (i.e. trajectory) use
-```julia
-Reconstruction(tr::SizedAray{A, B}, D, τ)
-Reconstruction(tr::AbstractDataset{B}, D, τ)
-```
-with `B` the "base" dimensions.
+*Notice* - The dimension of the returned dataset is `D+1`!
 
-If the trajectory is for example ``(x, y)``, then the ``n``th row is
+### Multiple Timeseries
+To make a reconstruction out of a multiple timeseries (i.e. trajectory) the number
+of timeseries must be known by type, so `s` can be either:
+
+    * `s::AbstractDataset{B}`
+    * `s::SizedAray{A, B}`
+
+If the trajectory is for example ``(x, y)`` and `τ` is integer, then the ``n``-th
+entry of the embedded space is
 ```math
-(x(n), y(n), x(n+\\tau), y(n+\\tau), \\dots, x(n+(D-1)\\tau), y(n+(D-1)\\tau))
+(x(n), y(n), x(n+\\tau), y(n+\\tau), \\dots, x(n+D\\tau), y(n+D\\tau))
 ```
-for integer `τ` and if `τ` is an `AbstractMatrix{Int}`, so that `size(τ) == (D, B)`,
-then the ``n``th row is
+If `τ` is an `AbstractMatrix{Int}`, so that `size(τ) == (D, B)`,
+then we have
 ```math
-(x(n+\\tau[1, 1]), y(n+\\tau[1, 2]), \\dots, x(n+\\tau[D, 1]), y(n+\\tau[D, 2]))
+(x(n), y(n), x(n+\\tau[1, 1]), y(n+\\tau[1, 2]), \\dots, x(n+\\tau[D, 1]), y(n+\\tau[D, 2]))
 ```
 
-Note that a reconstruction created
-this way will have `B*D` total dimensions and *not* `D`, as a result of
-each dimension of `s` having `D` delayed dimensions.
+*Notice* - The dimension of the returned dataset is `(D+1)*B`!
 
 ## References
-
 [1] : F. Takens, *Detecting Strange Attractors in Turbulence — Dynamical
 Systems and Turbulence*, Lecture Notes in Mathematics **366**, Springer (1981)
 
@@ -63,144 +106,76 @@ Systems and Turbulence*, Lecture Notes in Mathematics **366**, Springer (1981)
 
 [3] : K. Judd & A. Mees, [Physica D **120**, pp 273 (1998)](https://www.sciencedirect.com/science/article/pii/S0167278997001188)
 """
-struct Reconstruction{D, T<:Number, τ} <: AbstractReconstruction{D, T, τ}
-    data::Vector{SVector{D,T}}
-    delay::τ
-end
-
-function Reconstruction(s::AbstractVector{T}, D, τ::DT) where {T, DT}
-    if DT <: AbstractVector{Int}
-        length(τ) != D && throw(ArgumentError(
-        "The delay vector must have `length(τ) == D`."
-        ))
-    elseif DT != Int
-        throw(ArgumentError(
-        "Only Int or AbstractVector{Int} types are allowed for the delay."
-        ))
+@inline function reconstruct(s::AbstractVector{T}, D, τ) where {T}
+    de = DelayEmbedding(D, τ)
+    L = length(s) - de.delays[end]
+    data = Vector{SVector{D+1, T}}(undef, L)
+    @inbounds for i in 1:L
+        data[i] = de(s, i)
     end
-    Reconstruction{D, T, DT}(reconstruct(s, Val{D}(), τ), τ)
+    return Dataset{D+1, T}(data)
 end
 
-function reconstruct_impl(::Val{D}) where D
-    gens = [:(s[i + $k*τ]) for k=0:D-1]
-
-    quote
-        L = length(s) - ($(D-1))*τ;
-        T = eltype(s)
-        data = Vector{SVector{$D, T}}(L)
-        for i in 1:L
-            data[i] = SVector{$D,T}($(gens...))
-        end
-        V = typeof(s)
-        T = eltype(s)
-        data
-    end
-end
-function reconstruct_impl_tvec(::Val{D}) where D
-    gens = [:(s[i + τ[$k]]) for k=1:D]
-
-    quote
-        L = length(s) - ($(D-1))*maximum(τ);
-        T = eltype(s)
-        data = Vector{SVector{$D, T}}(L)
-        for i in 1:L
-            data[i] = SVector{$D,T}($(gens...))
-        end
-        V = typeof(s)
-        T = eltype(s)
-        data
-    end
-end
-@generated function reconstruct(s::AbstractVector{T}, ::Val{D}, τ::Int) where {D, T}
-    reconstruct_impl(Val{D}())
-end
-@generated function reconstruct(
-    s::AbstractVector{T}, ::Val{D}, τ::AbstractArray{Int}) where {D, T}
-    reconstruct_impl_tvec(Val{D}())
-end
-
-
-# Pretty print:
-Base.summary(d::Reconstruction{D, T, τ}) where {D, T, τ} =
-"(D=$(D), τ=$(d.delay)) - delay coordinates Reconstruction"
 #####################################################################################
 #                              MultiDimensional R                                   #
 #####################################################################################
-struct MDReconstruction{DxB, D, B, T<:Number, τ} <: AbstractReconstruction{DxB, T, τ}
-    data::Vector{SVector{DxB,T}}
-    delay::τ
+"""
+    MTDelayEmbedding(D, τ, B) -> `embedding`
+Return a delay coordinates embedding structure to be used as a functor,
+given multiple timeseries (`B` in total), either as a [`Dataset`](@ref) or a
+`SizedArray` (see [`reconstruct`](@ref)), and some index.
+Calling
+```julia
+embedding(s, n)
+```
+will create the `n`-th reconstructed vector of the embedded space, which has `D`
+temporal neighbors with delay(s) `τ`. See [`reconstruct`](@ref) for more.
+
+*Be very careful when choosing `n`, because `@inbounds` is used internally.*
+"""
+struct MTDelayEmbedding{D, B, X} <: AbstractEmbedding
+    # Again, here D is the number of temporal neighbors *plus one*.
+    delays::SMatrix{D, B, Int, X} # X = D*B = total dimension number
 end
 
-function reconstructmat_impl(::Val{S2}, ::Val{D}) where {S2, D}
-    gens = [:(s[i + $k*τ, $d]) for k=0:D-1 for d=1:S2]
-
-    quote
-        L = size(s,1) - ($(D-1))*τ;
-        T = eltype(s)
-        data = Vector{SVector{$D*$S2, T}}(L)
-        for i in 1:L
-            data[i] = SVector{$D*$S2,T}($(gens...))
-        end
-        V = typeof(s)
-        T = eltype(s)
-        data
+# notice that here D is the number of temporal neighbors WITHOUT +1 !
+function MTDelayEmbedding(D, τ, B)
+    X = (D+1)*B
+    if typeof(τ) <: Integer
+        idxs = SMatrix{D+1,B,Int,X}([k*τ for k in 0:D, j in 1:B])
+        return MTDelayEmbedding{D+1, B, X}(idxs)
+    elseif typeof(τ) <: AbstractMatrix{<:Integer}
+        D != size(τ)[1] && throw(ArgumentError(
+        "`size(τ)[1]` must equal the number of spatial neighbors."
+        ))
+        return MTDelayEmbedding{D+1, B, X}(SMatrix{D+1, B, Int, X}(
+        vcat(zeros(Int, B)', τ)))
+    else
+    	throw(ArgumentError("Please make sure τ is a Matrix or an Integer."))
     end
 end
 
-@generated function reconstruct(s::SizedArray{Tuple{A, B}, T, 2, M}, ::Val{D}, τ) where {A, B, T, M, D}
-    reconstructmat_impl(Val{B}(), Val{D}())
-end
-@generated function reconstruct(s::AbstractDataset{B, T}, ::Val{D}, τ) where {B, T, D}
-    reconstructmat_impl(Val{B}(), Val{D}())
-end
+@generated function (r::MTDelayEmbedding{D, B, X})(
+    s::Union{AbstractDataset{B, T}, SizedArray{Tuple{A, B}, T, 2, M}},
+    i) where {D, A, B, T, M, X}
 
-Reconstruction(s::AbstractDataset{B, T}, D, τ::Int) where {B, T} =
-MDReconstruction{B*D, D, B, T, Int}(reconstruct(s, Val{D}(), τ), τ)
-
-Reconstruction(s::SizedArray{Tuple{A, B}, T, 2, M}, D, τ::Int) where {A, B, T, M} =
-MDReconstruction{B*D, D, B, T, Int}(reconstruct(s, Val{D}(), τ), τ)
-
-## Multi-time version
-function reconstructmat_impl_tvec(::Val{S2}, ::Val{D}) where {S2, D}
-    gens = [:(s[i + τ[$k, $d], $d]) for k=1:D for d=1:S2]
+    gens = [:(s[i + r.delays[$k, $d], $d]) for k=1:D for d=1:B]
 
     quote
-        L = size(s,1) - maximum(τ);
-        T = eltype(s)
-        data = Vector{SVector{$D*$S2, T}}(L)
-        for i in 1:L
-            data[i] = SVector{$D*$S2,T}($(gens...))
-        end
-        V = typeof(s)
-        T = eltype(s)
-        data
+        @inbounds return SVector{$D*$B,T}($(gens...))
     end
 end
 
-@generated function reconstruct_multi(s::SizedArray{Tuple{A, B}, T, 2, M}, ::Val{D}, τ) where {A, B, T, M, D}
-    reconstructmat_impl_tvec(Val{B}(), Val{D}())
-end
-@generated function reconstruct_multi(s::AbstractDataset{B, T}, ::Val{D}, τ) where {B, T, D}
-    reconstructmat_impl_tvec(Val{B}(), Val{D}())
-end
+function reconstruct(
+    s::Union{AbstractDataset{B, T}, SizedArray{Tuple{A, B}, T, 2, M}},
+    D, τ) where {A, B, T, M}
 
-function Reconstruction(
-    s::AbstractDataset{B, T}, D, τ::DT) where {B, T, DT<:AbstractMatrix{Int}}
-    size(τ) != (D, B) && throw(ArgumentError(
-    "The delay matrix must have `size(τ) == (D, B)`."
-    ))
-    return MDReconstruction{B*D, D, B, T, DT}(reconstruct_multi(s, Val{D}(), τ), τ)
+    de = MTDelayEmbedding(D, τ, B)
+    L = size(s)[1] - maximum(de.delays)
+    X = (D+1)*B
+    data = Vector{SVector{X, T}}(undef, L)
+    @inbounds for i in 1:L
+        data[i] = de(s, i)
+    end
+    return Dataset{B*(D+1), T}(data)
 end
-
-function Reconstruction(
-    s::SizedArray{Tuple{A, B}, T, 2, M}, D, τ::DT
-    ) where {A, B, T, M, DT<:AbstractMatrix{Int}}
-    size(τ) != (D, B) && throw(ArgumentError(
-    "The delay matrix must have `size(τ) == (D, B)`."
-    ))
-    return MDReconstruction{B*D, D, B, T, DT}(reconstruct_multi(s, Val{D}(), τ), τ)
-end
-
-# Pretty print:
-Base.summary(d::MDReconstruction{DxB, D, B, T, τ}) where {DxB, D, B, T, τ} =
-"(B=$(B), D=$(D), τ=$(d.delay)) - delay coordinates multi-dimensional Reconstruction"